Rename `dqt/` to `src/`
