+++ main.py
@@ -9,9 +9,9 @@ if __name__ == '__main__':
             from utils.package_installer import install_dependencies
             install_dependencies()
             
-            from dqt.tracker import Tracker
+            from src.tracker import Tracker
             from settings import CONFIGS
-            from dqt.styletext import StyleText as Txt
+            from src.styletext import StyleText as Txt
         except ModuleNotFoundError as e:
             print("\n*!* —————————————————————————————— *!*")
             print(Txt("\n❌ Error!").bold().red())
